requirements.txt
aiohttp==0.21.4
aioTelegramBot==0.1
attrs==15.2.0
cffi==1.5.2
chardet==2.3.0
cryptography==1.2.3
idna==2.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pyOpenSSL==0.15.1
pyPlugin==0.1.2
python-dateutil==2.5.0
requests==2.9.1
service-identity==16.0.0
six==1.10.0
TelegramBotAPI==0.3.1
wolframalpha==2.3
